libcrt: add minimum pthread subset required for BoringSSL
Copy the minimum set of pthread functionality required for BoringSSL
from musl to libcrt. These are pthread_mutex_*(), pthread_once(),
pthread_rwlock_*() and the TLS accessors.

Trim their implementation as appropriate for the SVSM execution
environment:
- turn futex waits into spins (and make futex wakes a nop),
- in particular, turn timed waits into unbounded spins (because
  clock_gettime() is currently not reentrant),
- only PTHREAD_MUTEX_NORMAL is supported, everything else yields
  ENOSYS,
- for now, the TLS accessors are stubs always returning failure.

Note that the always failing TLS accessor stubs, i.e.
pthread_key_create(), pthread_setspecific() and pthread_getspecific()
will make BoringSSL to resort to initializing a new RNG on the stack for
every single OPENSSL_rand_bytes() request. The TLS primitives should
eventually get wired up to the per-CPU storage, but for now the
solution is good enough.

A note on the headers: unfortunately, BoringSSL cannot get built against
libcrt as it requires a C++ STL, so it needs to get built against the one
available on the host. However, the libcrt pthread type definitions must
still get injected for the BoringSSL structs to get layouted correctly.
Place the pthread.h in a include/pthread/ subirectory so that only that
header will be found when adding the path via -I to CPPFLAGS. In that
context note that the pthread.h contains much more declarations than what's
actually defined. That's needed because the host non-pthread libc headers
might still depend on them.

Signed-off-by: Nicolai Stange <nstange@suse.de>