Integrate cocoon-tpm-crypto
Integrate the cocoon-tpm-crypto crate into SVSM.

Introduce the "boringssl" Cargo feature for controlling the crypto
backend to be wired up behind the scenes: if enabled, BoringSSL is
used, crypto implementations in pure Rust otherwise. Default "boringssl"
to off for now.

For BoringSSL integration, define the bssl-bare-sys-target-integration
override package for tweaking the the build as appropriate for SVSM.
See [1] for details.

[1] https://github.com/nicstange/cocoon-tpm/tree/main/bssl-bare-sys

Signed-off-by: Nicolai Stange <nstange@suse.de>